LayoutTests/http/tests/workers/service/shownotification-denied-expected.txt

This tests that a notification shown by a service worker can be clicked and closed, with those actions being observable by the service worker

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "denied"
PASS Notification failed to show
